aarch64: kernel: fix get_processor_count return type
- The underlying field is u32
- The function for x86_64 returns a u32
- So we change this function to return a u32 as well

+++ src/arch/aarch64/kernel/mod.rs
@@ -104,8 +104,8 @@ pub fn get_mbinfo() -> usize {
 	unsafe { volatile_load(&BOOT_INFO.mb_info) as usize }
 }
 
-pub fn get_processor_count() -> usize {
-	unsafe { volatile_load(&BOOT_INFO.cpu_online) as usize }
+pub fn get_processor_count() -> u32 {
+	unsafe { volatile_load(&BOOT_INFO.cpu_online) }
 }
 
 /// Whether HermitCore is running under the "uhyve" hypervisor.
